multiple tuner support
CheckTunerAvailable() does a simple CURLOpen, and will fail if a http status code
of >= 400 is returned when connecting. This will then allow _GetChannelStreamURL()
to iterate through other tuners.

This should allow the case with multiple tuner devices that support the same channel
As the uniqueid is created with the url of the tuner, this wont match on a second tuner,
therefore check for a match of the ChannelNumber, SubChannel Number and ChannelName
are all equal to the requested channel