handle new and lost tuners
If a new tuner is found, add it to the list and process as normal
If total tuners found is less than the current list, clear and recreate tuner list for all
This doesnt not handle the situation of the same tuner count, but different tuners. It
will have a greater size on next update call, so should even out in the end, but it will go
an hour between the next update call before it is fixed.